+++ make/inc/customizer/google-fonts.php
@@ -7,7 +7,7 @@ if ( ! function_exists( 'ttfmake_get_google_fonts' ) ) :
 /**
  * Return an array of all available Google Fonts.
  *
- * Updated: 2015-01-12T22:51:47+00:00
+ * Updated: 2015-01-27T01:01:05+00:00
  *
  * @since  1.0.0.
  *
@@ -3074,6 +3074,17 @@ function ttfmake_get_google_fonts() {
 			),
 			'category' => 'sans-serif',
 		),
+		'Gurajada' => array(
+			'label' => 'Gurajada',
+			'variants' => array(
+				'regular',
+			),
+			'subsets' => array(
+				'latin',
+				'telugu',
+			),
+			'category' => 'serif',
+		),
 		'Habibi' => array(
 			'label' => 'Habibi',
 			'variants' => array(
@@ -3925,6 +3936,17 @@ function ttfmake_get_google_fonts() {
 				'latin-ext',
 			),
 			'category' => 'serif',
+		),
+		'Lakki Reddy' => array(
+			'label' => 'Lakki Reddy',
+			'variants' => array(
+				'regular',
+			),
+			'subsets' => array(
+				'latin',
+				'telugu',
+			),
+			'category' => 'handwriting',
 		),
 		'Lancelot' => array(
 			'label' => 'Lancelot',
@@ -6216,6 +6238,17 @@ function ttfmake_get_google_fonts() {
 			),
 			'category' => 'sans-serif',
 		),
+		'Ravi Prakash' => array(
+			'label' => 'Ravi Prakash',
+			'variants' => array(
+				'regular',
+			),
+			'subsets' => array(
+				'latin',
+				'telugu',
+			),
+			'category' => 'display',
+		),
 		'Redressed' => array(
 			'label' => 'Redressed',
 			'variants' => array(
@@ -7170,6 +7203,17 @@ function ttfmake_get_google_fonts() {
 			),
 			'category' => 'display',
 		),
+		'Sree Krushnadevaraya' => array(
+			'label' => 'Sree Krushnadevaraya',
+			'variants' => array(
+				'regular',
+			),
+			'subsets' => array(
+				'latin',
+				'telugu',
+			),
+			'category' => 'serif',
+		),
 		'Stalemate' => array(
 			'label' => 'Stalemate',
 			'variants' => array(
@@ -7277,6 +7321,28 @@ function ttfmake_get_google_fonts() {
 				'latin',
 			),
 			'category' => 'display',
+		),
+		'Suranna' => array(
+			'label' => 'Suranna',
+			'variants' => array(
+				'regular',
+			),
+			'subsets' => array(
+				'latin',
+				'telugu',
+			),
+			'category' => 'serif',
+		),
+		'Suravaram' => array(
+			'label' => 'Suravaram',
+			'variants' => array(
+				'regular',
+			),
+			'subsets' => array(
+				'latin',
+				'telugu',
+			),
+			'category' => 'serif',
 		),
 		'Suwannaphum' => array(
 			'label' => 'Suwannaphum',
@@ -7422,6 +7488,17 @@ function ttfmake_get_google_fonts() {
 				'latin',
 			),
 			'category' => 'serif',
+		),
+		'Timmana' => array(
+			'label' => 'Timmana',
+			'variants' => array(
+				'regular',
+			),
+			'subsets' => array(
+				'latin',
+				'telugu',
+			),
+			'category' => 'sans-serif',
 		),
 		'Tinos' => array(
 			'label' => 'Tinos',